+++ Lab07/image.h
@@ -18,7 +18,7 @@ public:
     ~Image();
     void operator=(const Image &rhs);
     void addLine(int x0, int y0, int x1, int y1, unsigned int color);
-    void addTriangle(int x0, int y0, int x1, int y1, int x2, int y2, unsigned int color);
+    void addTriangle(double x0, double y0, double x1, double y1, double x2, double y2, unsigned int color);
     void draw(GraphicsContext *gc, ViewContext *vc);
     void erase();
     Image undoShape(Image im);
