+++ Lab07/triangle.cpp
@@ -9,7 +9,7 @@
 using namespace std;
 
 // Triangle constructor
-Triangle::Triangle(int x0, int y0, int x1, int y1, int x2, int y2, unsigned int color)
+Triangle::Triangle(double x0, double y0, double x1, double y1, double x2, double y2, unsigned int color)
 {
     this->coord0[0][0] = x0;
     this->coord0[1][0] = y0;
