+++ Lab07/triangle.h
@@ -10,7 +10,7 @@ using namespace std;
 class Triangle : public Shape
 {
 public:
-    Triangle(int x0, int y0, int x1, int y1, int x2, int y2, unsigned int color);
+    Triangle(double x0, double y0, double x1, double y1, double x2, double y2, unsigned int color);
     Shape *clone();
     void draw(GraphicsContext *gc, ViewContext *vc);
